fixup! Add script for OnExitApplication(RESOURCE_CONSTRAINT)
